requirements.txt
Flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
gunicorn==21.2.0
pymongo==4.6.1
bcrypt==4.1.2
PyJWT==2.8.0
openpyxl==3.1.2
tzdata==2024.1
